feat:스무고개 데모데이 시연용 완성
