+++ frontend/front_twenty/twenty_guess.css
@@ -91,8 +91,8 @@ number{
 
 #back_botton{
     position: relative;
-    bottom : 10px;
-    right: 350px;
+    right: 430px;
+    top: 30px;
 }
 
 #reset_botton{
