+++ frontend/front_twenty/twenty_main.html
@@ -50,7 +50,7 @@
         </main_logo>
 
         <start_button>
-            <buttonype="button" class="image" id = "start_botton" onclick="next_guess();" onclick="location.href='./twenty_guess.html'";">
+            <buttonype="button" class="image" id = "start_botton" onclick="location.href='./twenty_guess.html'";">
             <img  src=".\img_file_twenty\twenty_main_sec_botton.svg" width="auto", height="70">
             </button>        
         </start_button>
